#767 moved job database insert into etl module
