+++ src/a18_etl_toolbox/test_/test_tran_sqlstrs_z_job.py
@@ -1,10 +1,4 @@
-from src.a00_data_toolbox.db_toolbox import (
-    db_table_exists,
-    get_create_table_sqlstr,
-    get_table_columns,
-    required_columns_exist,
-    create_select_query,
-)
+from src.a00_data_toolbox.db_toolbox import db_table_exists, get_create_table_sqlstr
 from src.a02_finance_logic._utils.strs_a02 import fisc_label_str, owner_name_str
 from src.a06_bud_logic._utils.str_a06 import (
     budunit_str,
@@ -22,7 +16,8 @@ from src.a06_bud_logic._utils.str_a06 import (
 from src.a10_bud_calc.bud_calc_config import get_bud_calc_config_dict
 from src.a17_idea_logic.idea_config import get_idea_sqlite_types
 from src.a17_idea_logic.idea_db_tool import get_default_sorted_list
-from src.a20_lobby_db_toolbox.lobby_sqlstrs import (
+from src.a18_etl_toolbox.tran_sqlstrs import (
+    create_prime_tablename as prime_table,
     get_job_create_table_sqlstrs,
     create_job_tables,
 )
@@ -41,7 +36,7 @@ def test_get_job_create_table_sqlstrs_ReturnsObj():
         # print(f"{x_dimen} checking...")
         x_config = bud_calc_config.get(x_dimen)
 
-        job_table = f"{x_dimen}_job"
+        job_table = prime_table(x_dimen, "job", None)
         job_cols = {fisc_label_str(), owner_name_str()}
         job_cols.update(set(x_config.get("jkeys").keys()))
         job_cols.update(set(x_config.get("jvalues").keys()))
@@ -63,17 +58,30 @@ def test_create_job_tables_CreatesTables():
         cursor.execute("SELECT COUNT(*) FROM sqlite_master WHERE type = 'table'")
         assert cursor.fetchone()[0] == 0
 
-        budmemb_job_table = f"{bud_acct_membership_str()}_job"
-        budacct_job_table = f"{bud_acctunit_str()}_job"
-        budgrou_job_table = f"{bud_groupunit_str()}_job"
-        budawar_job_table = f"{bud_concept_awardlink_str()}_job"
-        budfact_job_table = f"{bud_concept_factunit_str()}_job"
-        budheal_job_table = f"{bud_concept_healerlink_str()}_job"
-        budprem_job_table = f"{bud_concept_reason_premiseunit_str()}_job"
-        budares_job_table = f"{bud_concept_reasonunit_str()}_job"
-        budlabor_job_table = f"{bud_concept_laborlink_str()}_job"
-        budconc_job_table = f"{bud_conceptunit_str()}_job"
-        budunit_job_table = f"{budunit_str()}_job"
+        budmemb_job_table = prime_table(bud_acct_membership_str(), "job", None)
+        budacct_job_table = prime_table(bud_acctunit_str(), "job", None)
+        budgrou_job_table = prime_table(bud_groupunit_str(), "job", None)
+        budawar_job_table = prime_table(bud_concept_awardlink_str(), "job", None)
+        budfact_job_table = prime_table(bud_concept_factunit_str(), "job", None)
+        budheal_job_table = prime_table(bud_concept_healerlink_str(), "job", None)
+        budprem_job_table = prime_table(
+            bud_concept_reason_premiseunit_str(), "job", None
+        )
+        budares_job_table = prime_table(bud_concept_reasonunit_str(), "job", None)
+        budlabo_job_table = prime_table(bud_concept_laborlink_str(), "job", None)
+        budconc_job_table = prime_table(bud_conceptunit_str(), "job", None)
+        budunit_job_table = prime_table(budunit_str(), "job", None)
+        # budmemb_job_table = f"{bud_acct_membership_str()}_job"
+        # budacct_job_table = f"{bud_acctunit_str()}_job"
+        # budgrou_job_table = f"{bud_groupunit_str()}_job"
+        # budawar_job_table = f"{bud_concept_awardlink_str()}_job"
+        # budfact_job_table = f"{bud_concept_factunit_str()}_job"
+        # budheal_job_table = f"{bud_concept_healerlink_str()}_job"
+        # budprem_job_table = f"{bud_concept_reason_premiseunit_str()}_job"
+        # budares_job_table = f"{bud_concept_reasonunit_str()}_job"
+        # budlabo_job_table = f"{bud_concept_laborlink_str()}_job"
+        # budconc_job_table = f"{bud_conceptunit_str()}_job"
+        # budunit_job_table = f"{budunit_str()}_job"
 
         assert db_table_exists(cursor, budmemb_job_table) is False
         assert db_table_exists(cursor, budacct_job_table) is False
@@ -83,7 +91,7 @@ def test_create_job_tables_CreatesTables():
         assert db_table_exists(cursor, budheal_job_table) is False
         assert db_table_exists(cursor, budprem_job_table) is False
         assert db_table_exists(cursor, budares_job_table) is False
-        assert db_table_exists(cursor, budlabor_job_table) is False
+        assert db_table_exists(cursor, budlabo_job_table) is False
         assert db_table_exists(cursor, budconc_job_table) is False
         assert db_table_exists(cursor, budunit_job_table) is False
 
@@ -105,7 +113,7 @@ def test_create_job_tables_CreatesTables():
         assert db_table_exists(cursor, budheal_job_table)
         assert db_table_exists(cursor, budprem_job_table)
         assert db_table_exists(cursor, budares_job_table)
-        assert db_table_exists(cursor, budlabor_job_table)
+        assert db_table_exists(cursor, budlabo_job_table)
         assert db_table_exists(cursor, budconc_job_table)
         assert db_table_exists(cursor, budunit_job_table)
         cursor.execute("SELECT COUNT(*) FROM sqlite_master WHERE type = 'table'")
