+++ src/a18_etl_toolbox/tran_sqlstrs.py
@@ -65,8 +65,13 @@ def get_dimen_abbv7(dimen: str) -> str:
 
 
 def create_prime_tablename(
-    idea_dimen_or_abbv7: str, sound: str, stage: str, put_del: str = None
+    idea_dimen_or_abbv7: str, phase: str, stage: str, put_del: str = None
 ) -> str:
+    """
+    phase must be one: 's', 'v', 'job'
+    stage must be one: 'raw', 'aag', 'vld'
+    """
+
     abbv_references = {
         "FISCASH": "fisc_cashbook",
         "FISDEAL": "fisc_dealunit",
@@ -94,8 +99,10 @@ def create_prime_tablename(
     tablename = idea_dimen_or_abbv7
     if abbv_references.get(idea_dimen_or_abbv7.upper()):
         tablename = abbv_references.get(idea_dimen_or_abbv7.upper())
-    if sound in {"s", "v"}:
-        tablename = f"{tablename}_{sound}"
+    if phase in {"s", "v", "job"}:
+        tablename = f"{tablename}_{phase}"
+    if stage is None:
+        return tablename
 
     return f"{tablename}_{put_del}_{stage}" if put_del else f"{tablename}_{stage}"
 
@@ -1232,3 +1239,37 @@ FROM (
 ORDER BY fisc_label, owner_name, event_int, deal_time
 ;
 """
+
+
+CREATE_JOB_BUDMEMB_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_acct_membership_job (fisc_label TEXT, owner_name TEXT, acct_name TEXT, group_title TEXT, credit_vote REAL, debtit_vote REAL, _credor_pool REAL, _debtor_pool REAL, _fund_give REAL, _fund_take REAL, _fund_agenda_give REAL, _fund_agenda_take REAL, _fund_agenda_ratio_give REAL, _fund_agenda_ratio_take REAL)"""
+CREATE_JOB_BUDACCT_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_acctunit_job (fisc_label TEXT, owner_name TEXT, acct_name TEXT, credit_belief REAL, debtit_belief REAL, _credor_pool REAL, _debtor_pool REAL, _fund_give REAL, _fund_take REAL, _fund_agenda_give REAL, _fund_agenda_take REAL, _fund_agenda_ratio_give REAL, _fund_agenda_ratio_take REAL, _inallocable_debtit_belief REAL, _irrational_debtit_belief REAL)"""
+CREATE_JOB_BUDGROU_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_groupunit_job (fisc_label TEXT, owner_name TEXT, group_title TEXT, fund_coin REAL, bridge TEXT, _credor_pool REAL, _debtor_pool REAL, _fund_give REAL, _fund_take REAL, _fund_agenda_give REAL, _fund_agenda_take REAL)"""
+CREATE_JOB_BUDAWAR_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_awardlink_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, awardee_title TEXT, give_force REAL, take_force REAL, _fund_give REAL, _fund_take REAL)"""
+CREATE_JOB_BUDFACT_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_factunit_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, fcontext TEXT, fbranch TEXT, fopen REAL, fnigh REAL)"""
+CREATE_JOB_BUDHEAL_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_healerlink_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, healer_name TEXT)"""
+CREATE_JOB_BUDPREM_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_reason_premiseunit_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, rcontext TEXT, pbranch TEXT, pnigh REAL, popen REAL, pdivisor INTEGER, _task INTEGER, _status INTEGER)"""
+CREATE_JOB_BUDREAS_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_reasonunit_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, rcontext TEXT, rcontext_concept_active_requisite INTEGER, _task INTEGER, _status INTEGER, _rcontext_concept_active_value INTEGER)"""
+CREATE_JOB_BUDLABO_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_concept_laborlink_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, labor_title TEXT, _owner_name_labor INTEGER)"""
+CREATE_JOB_BUDCONC_SQLSTR = """CREATE TABLE IF NOT EXISTS bud_conceptunit_job (fisc_label TEXT, owner_name TEXT, concept_way TEXT, begin REAL, close REAL, addin REAL, numor INTEGER, denom INTEGER, morph INTEGER, gogo_want REAL, stop_want REAL, mass INTEGER, pledge INTEGER, problem_bool INTEGER, fund_coin REAL, _active INTEGER, _task INTEGER, _fund_onset REAL, _fund_cease REAL, _fund_ratio REAL, _gogo_calc REAL, _stop_calc REAL, _level INTEGER, _range_evaluated INTEGER, _descendant_pledge_count INTEGER, _healerlink_ratio REAL, _all_acct_cred INTEGER, _all_acct_debt INTEGER)"""
+CREATE_JOB_BUDUNIT_SQLSTR = """CREATE TABLE IF NOT EXISTS budunit_job (fisc_label TEXT, owner_name TEXT, credor_respect REAL, debtor_respect REAL, fund_pool REAL, max_tree_traverse INTEGER, tally INTEGER, fund_coin REAL, penny REAL, respect_bit REAL, _rational INTEGER, _keeps_justified INTEGER, _offtrack_fund REAL, _sum_healerlink_share REAL, _keeps_buildable INTEGER, _tree_traverse_count INTEGER)"""
+
+
+def get_job_create_table_sqlstrs() -> dict[str, str]:
+    return {
+        "bud_acct_membership_job": CREATE_JOB_BUDMEMB_SQLSTR,
+        "bud_acctunit_job": CREATE_JOB_BUDACCT_SQLSTR,
+        "bud_groupunit_job": CREATE_JOB_BUDGROU_SQLSTR,
+        "bud_concept_awardlink_job": CREATE_JOB_BUDAWAR_SQLSTR,
+        "bud_concept_factunit_job": CREATE_JOB_BUDFACT_SQLSTR,
+        "bud_concept_healerlink_job": CREATE_JOB_BUDHEAL_SQLSTR,
+        "bud_concept_reason_premiseunit_job": CREATE_JOB_BUDPREM_SQLSTR,
+        "bud_concept_reasonunit_job": CREATE_JOB_BUDREAS_SQLSTR,
+        "bud_concept_laborlink_job": CREATE_JOB_BUDLABO_SQLSTR,
+        "bud_conceptunit_job": CREATE_JOB_BUDCONC_SQLSTR,
+        "budunit_job": CREATE_JOB_BUDUNIT_SQLSTR,
+    }
+
+
+def create_job_tables(conn_or_cursor: sqlite3_Connection):
+    for create_table_sqlstr in get_job_create_table_sqlstrs().values():
+        conn_or_cursor.execute(create_table_sqlstr)
